fix 割引計算の見直し

+++ src/main/kotlin/domain/model/discount/rate/DiscountRate.kt
@@ -1,10 +1,12 @@
 package domain.model.discount.rate
 
+import java.math.BigDecimal
+
 /**
  * 割引適用率
  */
-enum class DiscountRate(val value: Int) {
-    NO_DISCOUNT(0),
-    RATE_10(10),
-    RATE_15(15);
+enum class DiscountRate(val value: Double) {
+    NO_DISCOUNT(0.0),
+    RATE_10(10.0),
+    RATE_15(15.0);
 }
